requirements.txt
Django==3.0.3
Pillow==7.0.0
django-ckeditor==5.9.0
